+++ hyperdbg/libhyperdbg/code/debugger/commands/debugging-commands/dt-struct.cpp
@@ -49,7 +49,7 @@ CommandDtHelp()
     ShowMessages("\t\te.g : dt nt!_EPROCESS $proc\n");
     ShowMessages("\t\te.g : dt nt!_KPROCESS @rax+@rbx+c0\n");
     ShowMessages("\t\te.g : !dt nt!_EPROCESS 1f0300\n");
-    ShowMessages("\t\te.g : dt nt!_MY_STRUCT 7ff00040 pid 1420\n");
+    ShowMessages("\t\te.g : dt MyModule!_MY_STRUCT 7ff00040 pid 1420\n");
     ShowMessages("\t\te.g : dt nt!_EPROCESS $proc inline all\n");
     ShowMessages("\t\te.g : dt nt!_EPROCESS fffff8077356f010 inline no\n");
 }